backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic[email]>=2.0.0
python-jose[cryptography]>=3.3.0
argon2-cffi>=23.1.0
cryptography>=42.0.0
reportlab>=4.0.0
networkx>=3.3
rapidfuzz>=3.0.0
pyyaml>=6.0
python-multipart>=0.0.9
requests>=2.31.0
